refactor: rename session-blocks.internal.ts to _session-blocks.ts
• Move session-blocks.internal.ts to _session-blocks.ts for better file organization

• Update imports in commands/blocks.ts and data-loader.ts
